float values set with

+++ Linear_Classifier/src/linear_classifier/Point.java
@@ -10,16 +10,16 @@ package linear_classifier;
  * @author Bashar Sader
  */
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public int val;
     public Point(){
-        x = 0;
-        y = 0;
+        x = 0.0;
+        y = 0.0;
         val = 0;
     }
 
-    public Point(int x, int y, int val) {
+    public Point(double x, double y, int val) {
        this.x = x; this.y = y; this.val = val;
     }
 }
